requirements.txt
lxml
telebot
pyTelegramBotAPI
flake8==2.5.4
mock==2.0.0
kik==1.2.0
Flask==0.11
pyyaml
pep8
nose
nose-cov
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
click==6.6
gunicorn==19.6.0
itsdangerous==0.24
requests==2.10.0
wsgiref==0.1.2
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
spotipy
pyopenssl
uwsgi
sqlalchemy
